+++ src/cmd/collection/delete_schema.go
@@ -117,11 +117,11 @@ func runCollectionDeleteSchema(cmd *cobra.Command, args []string) {
 		} else {
 			err = utils.DeleteWeaviateCollectionSchema(ctx, dbConfig, collectionName)
 		}
-	case config.VectorDBTypeSupabase:
+	case config.VectorDBTypeSupabase, config.VectorDBTypeSupabaseCloud:
 		// Supabase collections don't have separate schemas to delete
 		utils.PrintWarning("Supabase collections don't have separate schemas to delete")
 		return
-	case config.VectorDBTypeMongoDB:
+	case config.VectorDBTypeMongoDB, config.VectorDBTypeMongoDBCloud:
 		// MongoDB collections don't have separate schemas
 		utils.PrintWarning("MongoDB collections don't have separate schemas to delete")
 		return