+++ src/cmd/document/delete.go
@@ -121,9 +121,9 @@ func runDocumentDelete(cmd *cobra.Command, args []string) {
 	switch dbConfig.Type {
 	case config.VectorDBTypeCloud, config.VectorDBTypeLocal:
 		utils.DeleteWeaviateDocuments(ctx, dbConfig, collectionName, args[1:], metadataFilters, virtual, pattern, name)
-	case config.VectorDBTypeSupabase:
+	case config.VectorDBTypeSupabase, config.VectorDBTypeSupabaseCloud:
 		utils.DeleteDocuments(ctx, dbConfig, collectionName, args[1:], metadataFilters, virtual, pattern, name)
-	case config.VectorDBTypeMongoDB:
+	case config.VectorDBTypeMongoDB, config.VectorDBTypeMongoDBCloud:
 		utils.DeleteDocuments(ctx, dbConfig, collectionName, args[1:], metadataFilters, virtual, pattern, name)
 	case config.VectorDBTypeMilvusLocal, config.VectorDBTypeMilvusCloud:
 		utils.DeleteDocuments(ctx, dbConfig, collectionName, args[1:], metadataFilters, virtual, pattern, name)